requirements.txt
click==7.1.2
redis==3.5.3
kombu==5.0.2
hiredis==1.1.0
Django==3.1.7
celery==5.0.5
coreapi==2.3.3
ipython==7.20.0
openpyxl==3.0.6
Pygments==2.8.0
Markdown==3.3.3
whitenoise==5.2.0
django-silk==4.1.0
django-redis==4.12.1
django-ipware==3.0.2
django-filter==2.4.0
django-oss-storage==1.1.1
django-crispy-forms==1.11.1
django-cors-headers==3.7.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-timezone-field==4.1.1
django-mirage-field==1.1.7
django-hashid-field==3.1.3
djangorestframework==3.12.2
python-redis-lock==3.7.0
python-crontab==2.5.1
phonenumbers==8.12.18
user-agents==2.2.0
decorator==4.4.2
pendulum==2.1.2
Pillow==8.1.0
pyotp==2.6.0
PyYAML==5.4.1
qrcode==6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
billiard==3.6.3.0
drf-yasg==1.20.0
cryptography==3.2.1
pycryptodome==3.10.1
zbarlight==3.0
pytz==2020.4
mock==4.0.3
py==1.9.0

daphne==3.0.1
asgiref==3.3.1
channels==3.0.3
websockets==8.1
channels-redis==3.2.0
pyOpenSSL==20.0.1
packaging==20.9

oss2==2.14.0
wechatpy==1.8.14
# aliyun-mns-sdk==1.1.5
aliyun-python-sdk-core==2.13.30
aliyun-python-sdk-dybaseapi==1.0.6
aliyun-python-sdk-dysmsapi==2.1.1
aliyun-python-sdk-dyvmsapi==3.0.2
aliyun-python-sdk-green==3.6.3
aliyun-python-sdk-kms==2.14.0

supervisor==4.2.1
uWSGI==2.0.19.1
fabric==2.6.0

prompt-toolkit
ruamel.yaml.clib
ruamel.yaml
traitlets

amqp
arrow
emoji
msgpack
tornado
autobahn
docutils
sentry-sdk
setuptools
requests
sqlparse
certifi
iso8601
urllib3
txaio
attrs
pytz
cffi
idna
tqdm
toml
h2
py

# pip install --upgrade pip --no-cache-dir
# pip install -r requirements.txt --upgrade --no-cache-dir
# pip install --upgrade -r requirements.txt -i https://pypi.org/simple/ --default-timeout=5
# pip install --upgrade -r requirements.txt -i https://mirrors.aliyun.com/pypi/simple/ --default-timeout=5
# pip list --outdated --format=freeze | grep -v '^\-e' | cut -d = -f 1  | xargs -n1 pip install -U
